Change config file format to TOML. Change file structure.
