--- .tbls-sf-cli-meta.json
@@ -1,14 +1,0 @@
-{
-    "replacements": [
-        {
-            "def": "",
-            "prop": "",
-            "char": "$["
-        },
-        {
-            "def": "tables",
-            "prop": "",
-            "char": "$["
-        }
-    ]
-}
