+++ go.mod
@@ -1,3 +1,7 @@
 module github.com/shellyln/tbls-driver-sf-cli-meta
 
 go 1.24
+
+require github.com/shellyln/go-loose-json-parser v0.0.19
+
+require github.com/shellyln/takenoco v0.0.11 // indirect
