+++ pkg/driver/driver.go
@@ -5,42 +5,18 @@ import (
 	"os"
 )
 
-func ReadSalseforceMeta(baseDir string) (SalesforceMeta, error) {
-	var retval SalesforceMeta
-	var err error
-
-	retval.GlobalValueSets, err = readGlobalValueSetsMeta(baseDir)
-	if err != nil {
-		return retval, err
-	}
-
-	retval.RestrictionRules, err = readRestrictionRulesMeta(baseDir)
-	if err != nil {
-		return retval, err
-	}
-
-	retval.Flows, err = readFlowsMeta(baseDir)
-	if err != nil {
-		return retval, err
-	}
-
-	retval.ApexTriggers, err = readApexTriggers(baseDir)
-	if err != nil {
-		return retval, err
-	}
-
-	retval.SObjects, err = readObjectsMeta(baseDir, retval.GlobalValueSets)
-	if err != nil {
-		return retval, err
-	}
-
-	removeMissingRelations(retval.SObjects)
-
-	return retval, nil
-}
-
 func Run() error {
-	baseDir, err := makeBaseDir()
+	curDir, err := os.Getwd()
+	if err != nil {
+		return err
+	}
+
+	config, err := ReadConfig(curDir)
+	if err != nil {
+		return err
+	}
+
+	baseDir, err := makeBaseDir(curDir)
 	if err != nil {
 		return err
 	}
@@ -55,7 +31,7 @@ func Run() error {
 		return err
 	}
 
-	err = PostProcess(baseDir, schema)
+	err = PostProcess(config, baseDir, schema)
 	if err != nil {
 		return err
 	}
