+++ pkg/driver/typespp.go
@@ -4,6 +4,8 @@ import "strings"
 
 type PpDriverConfig struct {
 	PpReplacements []PpReplacement `json:"replacements"`
+	Include        *[]string       `json:"include"`
+	Exclude        *[]string       `json:"exclude"`
 }
 
 type PpReplacement struct {
@@ -11,4 +13,9 @@ type PpReplacement struct {
 	Prop     string `json:"prop"`
 	Char     string `json:"char"`
 	replacer *strings.Replacer
+}
+
+type strPair struct {
+	A string
+	B string
 }
